feat: add base ws config
