+++ .eslintrc.js
@@ -13,9 +13,8 @@ module.exports = {
   root: true,
   env: {
     node: true,
-    jest: true,
   },
-  ignorePatterns: ['.eslintrc.js'],
+  ignorePatterns: ['.eslintrc.js', '.index.js'],
   rules: {
     '@typescript-eslint/interface-name-prefix': 'off',
     '@typescript-eslint/explicit-function-return-type': 'off',
