+++ package.json
@@ -2,10 +2,10 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
+    "main": "index.ts",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
+        "start": "node ./index.ts",
+        "start:dev": "nodemon ./index.ts",
         "start-server:dev": "nodemon --exec ts-node ./src/node_server/index.ts",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
@@ -14,7 +14,8 @@
     "license": "ISC",
     "devDependencies": {
         "@types/node": "^20.11.20",
-        "@typescript-eslint/eslint-plugin": "^6.21.0",
+        "@types/ws": "^8.5.10",
+        "@typescript-eslint/eslint-plugin": "^7.0.2",
         "@typescript-eslint/parser": "^7.0.2",
         "eslint": "^8.56.0",
         "eslint-config-prettier": "^9.1.0",
