+++ tsconfig.json
@@ -18,7 +18,4 @@
           ]
       }
   },
-  "include": [
-      "src/**/*"
-  ]
 }
